cleanup code + comments
