+++ pylot.py
@@ -297,6 +297,13 @@ def add_debugging_component(graph, top_down_camera_setup, carla_op, camera_ops,
         carla_op
     )
 
+
+    kalman_op = []
+    if FLAGS.log_kalman:
+        kalman_op = [pylot.operator_creator.create_kalman_logger_op(graph)]
+        graph.connect([carla_op], kalman_op)
+
+
     # Add recording operators.
     pylot.operator_creator.add_recording_operators(
         graph,
@@ -402,12 +409,6 @@ def main(argv):
                             lane_det_ops,
                             bgr_camera_setup)
 
-
-    kalman_op = []
-    if FLAGS.log_kalman:
-        kalman_op = [pylot.operator_creator.create_kalman_logger_op(graph)]
-        graph.connect([carla_op], kalman_op)
-        graph.connect([agent_op], kalman_op)
 
     # Add planning operators.
     goal_location = (234.269989014, 59.3300170898, 39.4306259155)
